requirements.txt
Flask
mysql-connector-python
plotly
